pyproject.toml
[project]
name = "kpn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.135.2",
    "jsonlines>=4.0.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-docling>=2.0.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "llama-index>=0.14.19",
    "llama-index-vector-stores-faiss>=0.6.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.1.1",
    "requests>=2.33.0",
]
